add Container as an abstract superclass for Folder and Document
